requirements.txt
fastapi>=0.68.1
uvicorn>=0.15.0
sqlalchemy>=1.4.23
pydantic>=1.8.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
aiofiles>=0.7.0
python-dotenv>=0.19.0
google-auth>=2.3.3
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
google-api-python-client>=2.31.0
pandas>=1.3.3
openpyxl>=3.0.9